Search and Capture while walking to Poke Stop
When initial transferring, use the Release Criteria in config.json
