+++ pokemongo_bot/cell_workers/__init__.py
@@ -3,5 +3,5 @@
 from pokemon_catch_worker import PokemonCatchWorker
 from seen_fort_worker import SeenFortWorker
 from move_to_fort_worker import MoveToFortWorker
-from initial_transfer_worker import InitialTransferWorker
+from transfer_worker import TransferWorker
 from evolve_all_worker import EvolveAllWorker
